Enable mix and match of layouts
